Rearranges some tests in its own files
